+++ Plugins/ProcGen/Source/NexusProcGen/Public/NProcGenDeveloperOverlayWidget.h
@@ -3,7 +3,6 @@
 
 #pragma once
 
-#include "INWidgetMessageProvider.h"
 #include "NDeveloperOverlayWidget.h"
 #include "NProcGenDeveloperOverlayWidget.generated.h"
 
@@ -12,17 +11,12 @@ class UNProcGenOperation;
 class UNListView;
 
 UCLASS(ClassGroup = "NEXUS", DisplayName = "ProcGen Developer Overlay Widget", BlueprintType, Abstract)
-class NEXUSPROCGEN_API UNProcGenDeveloperOverlayWidget : public UNDeveloperOverlayWidget, public INWidgetMessageProvider
+class NEXUSPROCGEN_API UNProcGenDeveloperOverlayWidget : public UNDeveloperOverlayWidget
 {
 	GENERATED_BODY()
 	
 	virtual void NativeConstruct() override;
 	virtual void NativeDestruct() override;
-	
-	//~INMessageProvider interface
-	virtual bool HasProvidedMessage() const override;
-	virtual FString GetProvidedMessage() const override;
-	//~End of INMessageProvider interface
 	
 	void OnOperationStatusChanged(UNProcGenOperation* Operation, const ENProcGenOperationState NewState);
 	
@@ -31,6 +25,10 @@ protected:
 	TObjectPtr<UNListView> OperationsList;	
 	
 private:
+	
+	bool HasItems();
+	void OnChanges() { };
+	
 	FDelegateHandle OperationsStatusChangedDelegateHandle;
 	static bool ShouldShowOperation(const FName& OperationName);
 };