+++ Plugins/ProcGen/Source/NexusProcGen/Public/NProcGenSettings.h
@@ -53,9 +53,6 @@ public:
 		meta=(ToolTip="What is the size of the player's collider?"))
 	FVector PlayerSize = FVector(72.f, 184.f, 72.f);
 
-	UPROPERTY(Config, EditAnywhere, BlueprintReadOnly,  Category = "Developer Overlay", DisplayName="Widget")
-	TSubclassOf<UNProcGenDeveloperOverlayWidget> DeveloperOverlayWidget;
-	
 	UPROPERTY(Config, EditAnywhere, BlueprintReadOnly,  Category = "Debug", DisplayName="Proxy Material")
 	TSoftObjectPtr<UMaterial> ProxyMaterial;
 
@@ -64,18 +61,6 @@ private:
 	void ValidateSettings()
 	{
 		bool bNeedsSave = false;
-
-		if (!DeveloperOverlayWidget)
-		{
-			UClass* DefaultOverlayClass = FSoftClassPath(
-				TEXT("/NexusProcGen/WB_NProcGenDeveloperOverlay.WB_NProcGenDeveloperOverlay_C"))
-				.TryLoadClass<UNProcGenDeveloperOverlayWidget>();
-			if (DefaultOverlayClass != nullptr)
-			{
-				bNeedsSave = true;
-				DeveloperOverlayWidget = DefaultOverlayClass;
-			}
-		}
 
 		if (!ProxyMaterial)
 		{